Introduce KeychainTracker
to more safely provide keychain functionality on top SpkTracker
